backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
mysql-connector-python==8.3.0
python-dotenv==1.0.1
